Improved data ingestion

+++ data-persistence-service/src/main/java/com/intelligent_personal_trainer/data_persistence_service/entity/FitnessDataEntity.java
@@ -24,7 +24,7 @@ public class FitnessDataEntity {
     private String userId;
 
     private double averageHeartRate;
-    private int totalSteps;
+    private double totalSteps;
     private double totalDistance;
     private double totalCaloriesBurned;
