+++ docker/init.sql
@@ -2,7 +2,7 @@ CREATE TABLE IF NOT EXISTS fitness_data (
         id BIGSERIAL NOT NULL,
     user_id VARCHAR(255) NOT NULL,
     average_heart_rate DOUBLE PRECISION,
-    total_steps INTEGER,
+    total_steps DOUBLE PRECISION,
     total_distance DOUBLE PRECISION,
     total_calories_burned DOUBLE PRECISION,
     workout_list JSONB,
