+++ fitness-data-common/src/main/java/com/intelligent_personal_trainer/common/data/FitnessData.java
@@ -16,7 +16,7 @@ public class FitnessData {
     private Instant timestamp;
     private String userId;
     private double averageHeartRate;
-    private int totalSteps;
+    private double totalSteps;
     private double totalDistance;
     private double totalCaloriesBurned;
     private List<WorkoutData> workoutDataList;
